Extract the reports page into a separate JS file
